+++ src/vs/workbench/services/localization/electron-sandbox/languagePackService.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { ILanguagePackService } from 'vs/platform/languagePacks/common/languagePacks';
+import { ILanguagePackService } from 'vs/platform/languagePacks/common/languagePack';
 import { registerSharedProcessRemoteService } from 'vs/platform/ipc/electron-sandbox/services';
 
 registerSharedProcessRemoteService(ILanguagePackService, 'languagePacks', { supportsDelayedInstantiation: true });
